Organising and structuring the code into multiple files and sections. #23
