+++ client/index.html
@@ -40,6 +40,13 @@
   
 
   <script src="https://cdn.socket.io/4.5.4/socket.io.min.js"></script>
+  <script src="Agent.js">     </script>
+  <script src="Node.js">      </script>
+  <script src="Resource.js">  </script>
+  <script src="State.js">     </script>
+  <script src="UI.js">        </script>
+  <script src="EventListeners.js"></script>
+  
   <script src="main.js"></script>
 </body>
 </html>
