+++ client/style.css
@@ -2,6 +2,14 @@ body {
     margin: 0;
     overflow: hidden;
     font-family: Arial, sans-serif;
+
+    /* Disable text selection */
+    -webkit-touch-callout: none;
+    -webkit-user-select: none;
+    -khtml-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
 }
 
 canvas {
