+++ package.json
@@ -2,5 +2,8 @@
   "dependencies": {
     "express": "^4.21.2",
     "socket.io": "^4.8.1"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.3"
   }
 }
